Flask/requirements.txt
aniso8601==8.1.1
APScheduler==3.7.0
bcrypt==3.2.0
bidict==0.21.2
blinker==1.4
braintree==4.8.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coinbase-commerce==1.0.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.0.2
Flask-Limiter==1.4
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
Flask-SocketIO==5.0.1
Flask-WTF==0.14.3
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
limits==1.5.1
MarkupSafe==1.1.1
mongoengine==0.22.1
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
python-dateutil==2.8.1
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
stripe==2.56.0
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
